refactor: clean streaming
