+++ templates/types/streaming/fastapi/app/api/routers/chat.py
@@ -33,90 +33,53 @@ async def chat(
         messages = data.get_history_messages()
 
         doc_ids = data.get_chat_document_ids()
-        if len(doc_ids) > 0:
-            filters = MetadataFilters(
-                filters=[
-                    MetadataFilter(
-                        key="private",
-                        value="true",
-                        operator="!=",  # type: ignore
-                    ),
-                    MetadataFilter(
-                        key="doc_id",
-                        value=doc_ids,
-                        operator="in",  # type: ignore
-                    ),
-                ],
-                condition="or",  # type: ignore
-            )
-        else:
-            filters = MetadataFilters(
-                filters=[
-                    MetadataFilter(
-                        key="private",
-                        value="true",
-                        operator="!=",  # type: ignore
-                    ),
-                ]
-            )
+        filters = generate_filters(doc_ids)
         logger.info("Creating chat engine with filters", filters.dict())
         chat_engine = get_chat_engine(filters=filters)
 
         event_handler = EventCallbackHandler()
         chat_engine.callback_manager.handlers.append(event_handler)  # type: ignore
 
-        async def content_generator():
-            # Yield the text response
-            async def _chat_response_generator():
-                response = await chat_engine.astream_chat(
-                    last_message_content, messages
-                )
-                async for token in response.async_response_gen():
-                    yield VercelStreamResponse.convert_text(token)
-                # the text_generator is the leading stream, once it's finished, also finish the event stream
-                event_handler.is_done = True
+        response = await chat_engine.astream_chat(last_message_content, messages)
 
-                # Yield the source nodes
-                yield VercelStreamResponse.convert_data(
-                    {
-                        "type": "sources",
-                        "data": {
-                            "nodes": [
-                                SourceNodes.from_source_node(node).dict()
-                                for node in response.source_nodes
-                            ]
-                        },
-                    }
-                )
-
-            # Yield the events from the event handler
-            async def _event_generator():
-                async for event in event_handler.async_event_gen():
-                    event_response = event.to_response()
-                    if event_response is not None:
-                        yield VercelStreamResponse.convert_data(event_response)
-
-            combine = stream.merge(_chat_response_generator(), _event_generator())
-            is_stream_started = False
-            async with combine.stream() as streamer:
-                async for output in streamer:
-                    if not is_stream_started:
-                        is_stream_started = True
-                        # Stream a blank message to start the stream
-                        yield VercelStreamResponse.convert_text("")
-
-                    yield output
-
-                    if await request.is_disconnected():
-                        break
-
-        return VercelStreamResponse(content=content_generator())
+        return VercelStreamResponse(request, event_handler, response)
     except Exception as e:
         logger.exception("Error in chat engine", exc_info=True)
         raise HTTPException(
             status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
             detail=f"Error in chat engine: {e}",
         ) from e
+
+
+def generate_filters(doc_ids):
+    if len(doc_ids) > 0:
+        filters = MetadataFilters(
+            filters=[
+                MetadataFilter(
+                    key="private",
+                    value="true",
+                    operator="!=",  # type: ignore
+                ),
+                MetadataFilter(
+                    key="doc_id",
+                    value=doc_ids,
+                    operator="in",  # type: ignore
+                ),
+            ],
+            condition="or",  # type: ignore
+        )
+    else:
+        filters = MetadataFilters(
+            filters=[
+                MetadataFilter(
+                    key="private",
+                    value="true",
+                    operator="!=",  # type: ignore
+                ),
+            ]
+        )
+
+    return filters
 
 
 # non-streaming endpoint - delete if not needed
